+++ ui/src/pages/Devices/DeviceFlyout/deviceFlyoutConstants.js
